Typescriptify metadata index
Replaces `lib/metadata/index.js` and its hand-written `index.d.ts` with
a single `index.ts`, so the declarations are derived from the
implementation instead of being maintained alongside it.

The module now exports `Metadata` as a named export rather than
assigning it to `module.exports`, which is what the old `index.d.ts`
already declared. Export metadata declarations in main.js, so user can
access metadata types.
