+++ lib/metadata/index.ts
@@ -1,0 +1,287 @@
+"use strict";
+
+/**
+ * Module containing classes and fields related to metadata.
+ * @module metadata
+ */
+
+import { token } from "../..";
+import { Host } from "../host";
+import types = require("../types");
+import { ColumnInfo } from "../types/cql-utils";
+
+import { Udt } from "./user-defined-type";
+import { TableMetadata } from "./table-metadata";
+import { MaterializedView } from "./materialized-view";
+import { KeyspaceMetadata } from "./keyspace-metadata";
+import { SchemaFunction } from "./schema-function";
+import { Aggregate } from "./aggregate";
+import ClientState = require("./client-state");
+// TODO: Remove when lib/client-options.js is converted to Typescript.
+// @ts-ignore
+import { ClientOptions } from "../client-options";
+
+export { Aggregate } from "./aggregate";
+export { SchemaFunction } from "./schema-function";
+export { Index, IndexKind } from "./schema-index";
+export { KeyspaceMetadata, Strategy, StrategyKind } from "./keyspace-metadata";
+export { TableMetadata, ColumnMetadata, ColumnKind } from "./table-metadata";
+export { MaterializedView } from "./materialized-view";
+export { Udt, UdtField } from "./user-defined-type";
+export { ClientState };
+
+/**
+ * @const
+ * @private
+ */
+const _selectTraceSession =
+    "SELECT * FROM system_traces.sessions WHERE session_id=%s";
+/**
+ * @const
+ * @private
+ */
+const _selectTraceEvents =
+    "SELECT * FROM system_traces.events WHERE session_id=%s";
+/**
+ * @const
+ * @private
+ */
+const _selectSchemaVersionPeers = "SELECT schema_version FROM system.peers";
+/**
+ * @const
+ * @private
+ */
+const _selectSchemaVersionLocal = "SELECT schema_version FROM system.local";
+/**
+ * @const
+ * @private
+ */
+const _traceMaxAttemps = 5;
+/**
+ * @const
+ * @private
+ */
+const _traceAttemptDelay = 400;
+
+/**
+ * The trace session generated by Cassandra when query tracing is enabled for a query, stored in
+ * the `sessions` and `events` tables of the `system_traces` keyspace.
+ * @alias module:metadata~QueryTrace
+ */
+export interface QueryTrace {
+    requestType: string;
+    coordinator: types.InetAddress;
+    parameters: { [key: string]: any };
+    startedAt: number | types.Long;
+    duration: number;
+    clientAddress: string;
+    events: Array<{
+        id: types.Uuid;
+        activity: any;
+        source: any;
+        elapsed: any;
+        thread: any;
+    }>;
+}
+
+/**
+ * Represents cluster and schema information.
+ * The metadata class acts as a internal state of the driver.
+ */
+class Metadata {
+    /**
+     * Creates a new instance of {@link Metadata}.
+     */
+    constructor(_options: ClientOptions, _controlConnection: unknown) {}
+
+    /**
+     * Gets the keyspace metadata by name.
+     * @param name Name of the keyspace.
+     * @returns The keyspace metadata, or `null` if it does not exist.
+     */
+    getKeyspace(name: string): KeyspaceMetadata | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets all keyspace metadata.
+     * @returns A map of all keyspaces indexed by name.
+     */
+    getKeyspaces(): Map<string, KeyspaceMetadata> {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the host list representing the replicas that contain the given partition key, token or token range.
+     *
+     * It uses the pre-loaded keyspace metadata to retrieve the replicas for a token for a given keyspace.
+     * When the keyspace metadata has not been loaded, it returns null.
+     * @param keyspaceName Name of the keyspace.
+     * @param token Can be Buffer (serialized partition key), Token or TokenRange.
+     */
+    getReplicas(
+        keyspaceName: string,
+        token: Buffer | token.Token | token.TokenRange,
+    ): Host[] {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the token ranges that define data distribution in the ring.
+     * @returns The ranges of the ring or empty set if schema metadata is not enabled.
+     */
+    getTokenRanges(): Set<token.TokenRange> {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the token ranges that are replicated on the given host, for the given keyspace.
+     * @param keyspaceName The name of the keyspace to get ranges for.
+     * @param host The host.
+     * @returns Ranges for the keyspace on this host or null if keyspace
+     * isn't found or hasn't been loaded.
+     */
+    getTokenRangesForHost(
+        keyspaceName: string,
+        host: Host,
+    ): Set<token.TokenRange> | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Constructs a Token from the input buffer(s) or string input. If a string is passed in
+     * it is assumed this matches the token representation reported by cassandra.
+     * @param components The token components.
+     * @returns constructed token from the input buffer.
+     */
+    newToken(components: Buffer[] | Buffer | string): token.Token {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Constructs a TokenRange from the given start and end tokens.
+     * @param start The start token.
+     * @param end The end token.
+     * @returns build range spanning from start (exclusive) to end (inclusive).
+     */
+    newTokenRange(start: token.Token, end: token.Token): token.TokenRange {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the definition of an user-defined type.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the UDT.
+     * @returns The UDT definition, or `null` if it does not exist.
+     */
+    getUdt(keyspaceName: string, name: string): Udt | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the definition of a table.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the Table.
+     * @returns The table metadata, or `null` if it does not exist.
+     */
+    getTable(keyspaceName: string, name: string): TableMetadata | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the definition of CQL functions for a given name.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the Function.
+     * @returns An array of schema function metadata.
+     */
+    getFunctions(keyspaceName: string, name: string): SchemaFunction[] {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets a definition of CQL function for a given name and signature.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the Function.
+     * @param signature Array of types of the parameters.
+     * @returns The schema function metadata, or `null` if it does not exist.
+     */
+    getFunction(
+        keyspaceName: string,
+        name: string,
+        signature: string[] | ColumnInfo[],
+    ): SchemaFunction | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the definition of CQL aggregate for a given name.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the aggregate.
+     * @returns An array of schema aggregate metadata.
+     */
+    getAggregates(keyspaceName: string, name: string): Aggregate[] {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets a definition of CQL aggregate for a given name and signature.
+     * @param keyspaceName Name of the keyspace.
+     * @param name Name of the aggregate.
+     * @param signature Array of types of the parameters.
+     * @returns The schema aggregate metadata, or `null` if it does not exist.
+     */
+    getAggregate(
+        keyspaceName: string,
+        name: string,
+        signature: string[] | ColumnInfo[],
+    ): Aggregate | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the definition of a CQL materialized view for a given name.
+     *
+     * Note that, unlike the rest of the {@link Metadata} methods, this method does not cache the result for following
+     * calls, as the current version of the Cassandra native protocol does not support schema change events for
+     * materialized views. Each call to this method will produce one or more queries to the cluster.
+     * @param keyspaceName Name of the keyspace
+     * @param name Name of the materialized view
+     * @returns The materialized view definition, or `null` if it does not exist.
+     */
+    getMaterializedView(
+        keyspaceName: string,
+        name: string,
+    ): MaterializedView | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Gets the trace session generated by Cassandra when query tracing is enabled for the
+     * query. The trace itself is stored in Cassandra in the `sessions` and
+     * `events` table in the `system_traces` keyspace and can be
+     * retrieve manually using the trace identifier.
+     * @param traceId Identifier of the trace session.
+     * @param consistency The consistency level to obtain the trace.
+     * @returns The trace session, or `null` if it does not exist.
+     */
+    getTrace(
+        traceId: types.Uuid,
+        consistency?: types.consistencies,
+    ): QueryTrace | null {
+        throw new Error("TODO: Not implemented");
+    }
+
+    /**
+     * Checks whether hosts that are currently up agree on the schema definition.
+     *
+     * This method performs a one-time check only, without any form of retry; therefore
+     * `protocolOptions.maxSchemaAgreementWaitSeconds` setting does not apply in this case.
+     * @returns `true` when all hosts agree on the schema and `false` when there is no agreement or when
+     * the check could not be performed (for example, if the control connection is down).
+     */
+    checkSchemaAgreement(): boolean {
+        throw new Error("TODO: Not implemented");
+    }
+}
+
+export { Metadata };
